# Add plugin that helps with performance benchmarking