--- a/docs/guides/index.rst
+++ b/docs/guides/index.rst
@@ -10,3 +10,4 @@ This section aims to provide hands-on guides to demonstrate how to use Trinity. 
    quickstart
    architecture
    writing_plugins
+   measuring_runtime_sync_performance