### What was wrong?

We don't have good tooling to measure Trinitys run time performance. For things that aren't easily quantifiable in isolation (e.g. with `timeit`), our best option so far is to just run Trinity and manually take some numbers. This PR tries to improve that a little.

### How was it fixed?

1. The `PerformanceTrackerPlugin` roughly does the following:
1.1 Measure the sync time from boot up to some target block
1.2 Also measure the time from the start of sync up to the target block (in other words don't include the upfront time until syncing has startet)
1.3 Write report to log and append it to the `performance_report.txt` file
1.4 Shutdown Trinity

2. The `TrackerRunnerPlugin` plugin is a `BaseMainProcessPlugin` plugin is a minimal viable runner that runs Trinity up to `n` runs with performance tracking enabled.
2.1 E.g. `trinity track-perf --track-perf-target-block 100000 --trag-perf-run-count 5`
2.2 After all runs are finished one can look at the results in `performance_report.txt` which should look like this:

```
Performance tracking finished run: Unnamed run #1
Hit target block 100
Total duration: 15s
Actual duration: 0s
#####################
Performance tracking finished run: Unnamed run #2
Hit target block 100
Total duration: 15s
Actual duration: 5s
#####################
Performance tracking finished run: Unnamed run #3
Hit target block 100
Total duration: 20s
Actual duration: 5s
#####################
```

With the default configuration, the tracker looks up the current head every 5 seconds which means that it won't be accurate to the very second. However, the general idea is that this should make it very easy to e.g. run Trinity ten times in a row up to block `1m` or something and hence give one a rough measure about its performance.

-  [ ] add changelog entry

[//]: # (For important changes, please add a new entry to the running release notes PR)
[//]: # (You can find the current one using: https://github.com/ethereum/trinity/pulls?q=is%3Aopen+is%3Apr+label%3A%22Release+Notes%22 )
[//]: # (See: https://trinity-client.readthedocs.io/en/latest/contributing.html#pull-requests)

#### Cute Animal Picture

![put a cute animal picture link inside the parentheses](https://static.independent.co.uk/s3fs-public/thumbnails/image/2016/09/05/10/cubs9.jpg?w968h681)