@cburgdorf 2019-05-06
Just dumping another idea here. We could add flags such as `--bin-a /path/to/trinity-a --bin-b path/to/trinity-b` so that the runner alternates between launching Trinity from two different locations, essentially A/B testing them. This might be a bit more accurate compared to first completing 10 runs on version x and then completing 10 runs on version y. Because it might be that network conditions have changed in between.