@cburgdorf 2019-05-10
A/B testing (as in continously toggling between two different programs/branches) is now implemented. One thing that would be super cool is to automatically output graphics such as those seen here https://www.parity.io/performance-analysis/