@cburgdorf 2019-06-27
@pipermerriam I added a bit of docs to this today. The code isn't really sophisticated but it's also 100 % plugin code that doesn't affect anything else in trinity and I think having this in `master` would be helpful (otherwise, testing branches does always mean I have to cherry-pick this in first).